requirements.txt
streamlit==1.31.1
pillow==10.2.0
google-generativeai==0.5.2
